requirements.txt
numpy      1.26.0
Pillow     10.0.1
pip        22.3
setuptools 65.5.0